Fixed some gamebreaking bugs in cross
One was an line that skipped to a certain portion of the code. Most likely for testing purposes.

Another was basically an order of evaluation issue. The label that gave you the correct objection was put after the label that gave you incorrect objections meaning that you could never escape since the game always found the "incorrect" label first.

I put some sound effects in to give player feedback when they are penalized.

I also changed the CROSS EXAMINATION text to match how the game does it.